Reorganize: clean directory structure
uccl_gin/        public device API (handle, rail, resources) — DeepEP include target
transport/       internal D2H+proxy+EFA layer (copied from ep/)
context.{hpp,cpp} host-side bridge between API and transport
bindings.cpp     Python C extension
tests/           flat: microbench.cu + all Python tests
python/          package helpers

Previously flat include/ and src/ with 20+ files mixed together.
