+++ experimental/uccl_gin/tests/test_microbench.py
@@ -3,7 +3,7 @@ import sys
 import os
 
 
-_ROOT = pathlib.Path(__file__).resolve().parents[2]
+_ROOT = pathlib.Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(_ROOT / "python"))
 
 from uccl_gin.microbench import MicrobenchConfig, run_microbench  # noqa: E402
